Rename gradebook.ts into gradebookTypes.ts for avoid confusion
